+++ packages/pixels/CHANGELOG.md
@@ -1,5 +1,26 @@
 # Changelog
 
+## [0.43.6](https://github.com/fuf-stack/pixels/compare/pixels-v0.43.5...pixels-v0.43.6) (2025-04-30)
+
+
+### Bug Fixes
+
+* **pixels:** test if tooltip size works without applying to base slot ([0b48b62](https://github.com/fuf-stack/pixels/commit/0b48b6263555b5bc5376ca64ad3738d27b79c243))
+
+## [0.43.5](https://github.com/fuf-stack/pixels/compare/pixels-v0.43.4...pixels-v0.43.5) (2025-04-30)
+
+
+### Bug Fixes
+
+* **pixels:** tooltip content overflow ([ec1c234](https://github.com/fuf-stack/pixels/commit/ec1c234fcf75e2867dca3b97431657f1838c4849))
+
+## [0.43.4](https://github.com/fuf-stack/pixels/compare/pixels-v0.43.3...pixels-v0.43.4) (2025-04-30)
+
+
+### Bug Fixes
+
+* **pixels:** add size prop to Tooltip ([fef27e6](https://github.com/fuf-stack/pixels/commit/fef27e620e4035378c3b2e1a05581f143b5420a2))
+
 ## [0.43.3](https://github.com/fuf-stack/pixels/compare/pixels-v0.43.2...pixels-v0.43.3) (2025-04-28)